LayoutTests/fast/canvas/webgl/texture-transparent-pixels-initialized-expected.txt

Tests there is no garbage in transparent regions of images uploaded as textures

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Checking lower left corner
PASS buf[idx] == 0 && buf[idx + 1] == 0 && buf[idx + 2] == 0 is true
PASS buf[idx] == 0 && buf[idx + 1] == 0 && buf[idx + 2] == 0 is true
PASS buf[idx] == 0 && buf[idx + 1] == 0 && buf[idx + 2] == 0 is true
PASS buf[idx] == 0 && buf[idx + 1] == 0 && buf[idx + 2] == 0 is true
Checking upper left corner
PASS buf[idx] == 0 && buf[idx + 1] == 0 && buf[idx + 2] == 0 is true
PASS buf[idx] == 0 && buf[idx + 1] == 0 && buf[idx + 2] == 0 is true
PASS buf[idx] == 0 && buf[idx + 1] == 0 && buf[idx + 2] == 0 is true
PASS buf[idx] == 0 && buf[idx + 1] == 0 && buf[idx + 2] == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
